# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sanctuary – anonymous safe social platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Dimension {
  id: string;
  name: string;
  nameAr: string;
  tagline: string;
  color: string;
  activeSpecters: number;
  icon: string;
  /** @nullable */
  description?: string | null;
}

export interface FeedItem {
  id: string;
  dimensionId: string;
  type: string;
  /** @nullable */
  content?: string | null;
  pulseCount?: number;
  createdAt: string;
}

export interface FeedPage {
  items: FeedItem[];
  /** @nullable */
  nextCursor: string | null;
}

export interface Galaxy {
  id: string;
  name: string;
  category: string;
  memberCount: number;
  color: string;
  description: string;
  isLive?: boolean;
}

export interface GalaxyInput {
  name: string;
  category: string;
  description: string;
  /** @nullable */
  color?: string | null;
}

export interface GalaxyMessage {
  id: string;
  galaxyId: string;
  content: string;
  spectralColor: string;
  createdAt: string;
}

export interface GalaxyMessagesPage {
  messages: GalaxyMessage[];
}

export interface GalaxyMessageInput {
  content: string;
  /** @nullable */
  spectralColor?: string | null;
}

export interface PollOption {
  id: string;
  text: string;
  votes: number;
}

export interface Poll {
  id: string;
  galaxyId: string;
  question: string;
  options: PollOption[];
  totalVotes: number;
  /** @nullable */
  endsAt?: string | null;
  createdAt: string;
}

export interface PollInput {
  question: string;
  options: string[];
  /** @nullable */
  endsInHours?: number | null;
}

export interface PollVoteInput {
  optionId: string;
}

export interface Tale {
  id: string;
  title: string;
  content: string;
  genre: string;
  pulseCount: number;
  commentCount: number;
  /** @nullable */
  excerpt?: string | null;
  createdAt: string;
}

export interface TaleInput {
  title: string;
  content: string;
  genre: string;
}

export interface TalesPage {
  items: Tale[];
  /** @nullable */
  nextCursor: string | null;
}

export interface Comment {
  id: string;
  taleId: string;
  content: string;
  spectralColor: string;
  createdAt: string;
}

export interface CommentInput {
  content: string;
  /** @nullable */
  spectralColor?: string | null;
}

export interface Reel {
  id: string;
  thumbnailUrl: string;
  /** @nullable */
  videoUrl?: string | null;
  emotionTag: string;
  resonanceCount: number;
  durationSeconds: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface ReelsPage {
  items: Reel[];
  /** @nullable */
  nextCursor: string | null;
}

export interface PulseResult {
  newCount: number;
}

export interface SecretInput {
  content: string;
}

export interface SecretResult {
  sent: boolean;
  message: string;
}

export interface ReceivedSecret {
  id: string;
  content: string;
  receivedAt: string;
}

export interface EchoSessionInput {
  spectralColor: string;
  /** @nullable */
  openingThought?: string | null;
}

export interface EchoSession {
  sessionId: string;
  status: string;
  partnerSpectralColor: string;
  createdAt: string;
}

export interface EchoMessage {
  id: string;
  sessionId: string;
  content: string;
  isMine: boolean;
  spectralColor: string;
  createdAt: string;
}

export interface EchoMessageInput {
  content: string;
}

export interface BreathingExercise {
  name: string;
  inhaleSeconds: number;
  holdSeconds: number;
  exhaleSeconds: number;
  description: string;
}

export interface Hotline {
  country: string;
  name: string;
  number: string;
  /** @nullable */
  available?: string | null;
}

export interface SanctuaryResources {
  breathingExercises: BreathingExercise[];
  hotlines: Hotline[];
  affirmations: string[];
}

export interface CheckInInput {
  mood: number;
  /** @nullable */
  note?: string | null;
}

export interface CheckInResponse {
  message: string;
  suggestion: string;
  resourcesAvailable?: boolean;
}

export interface DimensionStat {
  dimensionId: string;
  activeSpecters: number;
  recentActivity: number;
}

export interface AtriumStats {
  totalSpectersOnline: number;
  activeDimensions: number;
  dimensionStats: DimensionStat[];
}

export interface MoodEntry {
  id: string;
  mood: number;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  emotion?: string | null;
  createdAt: string;
}

export interface MoodInput {
  mood: number;
  spectralToken: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  emotion?: string | null;
}

export interface Bookmark {
  id: string;
  contentType: string;
  contentId: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  excerpt?: string | null;
  createdAt: string;
}

export interface BookmarkInput {
  contentType: string;
  contentId: string;
  spectralToken: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  excerpt?: string | null;
}

export interface DeleteResult {
  deleted: boolean;
}

export interface Expression {
  id: string;
  type: string;
  title: string;
  /** @nullable */
  description?: string | null;
  color: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  spectralAuthor?: string | null;
  resonanceCount: number;
  /** @nullable */
  size?: string | null;
  createdAt: string;
}

export interface SearchResults {
  tales: Tale[];
  galaxies: Galaxy[];
  expressions: Expression[];
  total: number;
}

export interface ExpressionInput {
  type: string;
  title: string;
  /** @nullable */
  description?: string | null;
  color: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  size?: string | null;
}

export interface ExpressionsPage {
  items: Expression[];
  /** @nullable */
  nextCursor: string | null;
}

export type ListGalaxiesParams = {
/**
 * @nullable
 */
category?: string | null;
};

export type ListTalesParams = {
/**
 * @nullable
 */
genre?: string | null;
/**
 * @nullable
 */
cursor?: string | null;
};

export type ListReelsParams = {
/**
 * @nullable
 */
cursor?: string | null;
};

export type ListMoodsParams = {
spectralToken: string;
/**
 * @nullable
 */
limit?: number | null;
};

export type ListBookmarksParams = {
spectralToken: string;
};

export type SearchContentParams = {
q: string;
/**
 * @nullable
 */
dimension?: string | null;
};

export type ListExpressionsParams = {
/**
 * @nullable
 */
type?: string | null;
/**
 * @nullable
 */
cursor?: string | null;
};