# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sanctuary – anonymous safe social platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /galaxies:
    get:
      operationId: listGalaxies
      tags: [galaxies]
      summary: List available galaxy spaces
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Galaxies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Galaxy"
    post:
      operationId: createGalaxy
      tags: [galaxies]
      summary: Create a new galaxy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalaxyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Galaxy"
  /galaxies/{galaxyId}/messages:
    get:
      operationId: getGalaxyMessages
      tags: [galaxies]
      summary: Get messages in a galaxy
      parameters:
        - name: galaxyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalaxyMessagesPage"
    post:
      operationId: sendGalaxyMessage
      tags: [galaxies]
      summary: Send an anonymous message to a galaxy
      parameters:
        - name: galaxyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalaxyMessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalaxyMessage"
...
  /tales/{taleId}/comments:
    get:
      operationId: getTaleComments
      tags: [comments]
      summary: Get comments on a tale
      parameters:
        - name: taleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addTaleComment
      tags: [comments]
      summary: Add an anonymous comment to a tale
      parameters:
        - name: taleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /messages/echo:
    post:
      operationId: startEchoSession
      tags: [messages]
      summary: Start a deep 1-1 echo conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EchoSessionInput"
      responses:
        "200":
          description: Echo session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EchoSession"
  /messages/echo/{sessionId}:
    get:
      operationId: getEchoMessages
      tags: [messages]
      summary: Get messages from an echo session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EchoMessage"
    post:
      operationId: sendEchoMessage
      tags: [messages]
      summary: Send a message in an echo session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EchoMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EchoMessage"
...
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: Get bookmarks for a spectral token
      parameters:
        - name: spectralToken
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: addBookmark
      tags: [bookmarks]
      summary: Bookmark a piece of content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Bookmarked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{bookmarkId}:
    delete:
      operationId: removeBookmark
      tags: [bookmarks]
      summary: Remove a bookmark
      parameters:
        - name: bookmarkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
  /expressions:
    get:
      operationId: listExpressions
      tags: [expressions]
      summary: List expression posts (art, poetry, motion)
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: cursor
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Expressions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpressionsPage"
    post:
      operationId: createExpression
      tags: [expressions]
      summary: Submit a new expression (art, poetry, visual, motion)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpressionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expression"
  /expressions/{expressionId}/resonate:
    post:
      operationId: resonateExpression
      tags: [expressions]
      summary: Resonate with an expression
      parameters:
        - name: expressionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resonated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PulseResult"
components:
  schemas:
...
    Galaxy:
      type: object
      required: [id, name, category, memberCount, color, description]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        memberCount:
          type: integer
        color:
          type: string
        description:
          type: string
        isLive:
          type: boolean
    GalaxyInput:
      type: object
      required: [name, category, description]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        color:
          type: ["string", "null"]
    GalaxyMessage:
      type: object
      required: [id, galaxyId, content, spectralColor, createdAt]
      properties:
        id:
          type: string
        galaxyId:
          type: string
        content:
          type: string
        spectralColor:
          type: string
        createdAt:
          type: string
    GalaxyMessagesPage:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GalaxyMessage"
    GalaxyMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        spectralColor:
          type: ["string", "null"]
...
    Comment:
      type: object
      required: [id, taleId, content, spectralColor, createdAt]
      properties:
        id:
          type: string
        taleId:
          type: string
        content:
          type: string
        spectralColor:
          type: string
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        spectralColor:
          type: ["string", "null"]
...
    PulseResult:
      type: object
      required: [newCount]
      properties:
        newCount:
          type: integer
...
    EchoSessionInput:
      type: object
      required: [spectralColor]
      properties:
        spectralColor:
          type: string
        openingThought:
          type: ["string", "null"]
    EchoSession:
      type: object
      required: [sessionId, status, partnerSpectralColor, createdAt]
      properties:
        sessionId:
          type: string
        status:
          type: string
        partnerSpectralColor:
          type: string
        createdAt:
          type: string
    EchoMessage:
      type: object
      required: [id, sessionId, content, isMine, spectralColor, createdAt]
      properties:
        id:
          type: string
        sessionId:
          type: string
        content:
          type: string
        isMine:
          type: boolean
        spectralColor:
          type: string
        createdAt:
          type: string
    EchoMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
...
    Bookmark:
      type: object
      required: [id, contentType, contentId, createdAt]
      properties:
        id:
          type: string
        contentType:
          type: string
        contentId:
          type: string
        title:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
        createdAt:
          type: string
    BookmarkInput:
      type: object
      required: [contentType, contentId, spectralToken]
      properties:
        contentType:
          type: string
        contentId:
          type: string
        spectralToken:
          type: string
        title:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
    DeleteResult:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean
...
    Expression:
      type: object
      required: [id, type, title, color, resonanceCount, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        imageUrl:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        spectralAuthor:
          type: ["string", "null"]
        resonanceCount:
          type: integer
        size:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpressionInput:
      type: object
      required: [type, title, color]
      properties:
        type:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        imageUrl:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
    ExpressionsPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Expression"
        nextCursor:
          type: ["string", "null"]