# lib/api-zod/src/generated/types/echoSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sanctuary – anonymous safe social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EchoSession {
  sessionId: string;
  status: string;
  partnerSpectralColor: string;
  createdAt: string;
}