# lib/api-zod/src/generated/types/echoSessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sanctuary – anonymous safe social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EchoSessionInput {
  spectralColor: string;
  /** @nullable */
  openingThought?: string | null;
}