# lib/api-zod/src/generated/types/expression.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sanctuary – anonymous safe social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Expression {
  id: string;
  type: string;
  title: string;
  /** @nullable */
  description?: string | null;
  color: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  spectralAuthor?: string | null;
  resonanceCount: number;
  /** @nullable */
  size?: string | null;
  createdAt: string;
}