# lib/api-zod/src/generated/types/galaxyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sanctuary – anonymous safe social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GalaxyInput {
  name: string;
  category: string;
  description: string;
  /** @nullable */
  color?: string | null;
}